+++ CHANGES.rst
@@ -6,7 +6,8 @@ There's a frood who really knows where his towel is.
 1.0a3 (unreleased)
 ------------------
 
-- Nothing changed yet.
+- Fix exception on ``AMPView`` when content has an Archetypes based lead image.
+  [hvelarde]
 
 
 1.0a2 (2016-10-06)